[Refactor] move GraduateDetails file path
